refactor(xds/generator): move route handling to package
